requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.1
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
